[editor] Externalize the SQL references repositories
To solve an issue with dynamic imports and web components we can now pass the UDF and SQL reference repositories as properties on the Editor component.

This also splits the previous repository in two, one for UDFs and one for the rest.
